Implement SSR for SlideOutView
