+++ testing/tests/DevExpress.serverSide/slideOutView.tests.js
@@ -1,0 +1,3 @@
+"use strict";
+
+require("../DevExpress.ui.widgets/slideOutView.markup.tests.js");
