+++ testing/tests/DevExpress.serverSide/widgetsCreation.tests.js
@@ -105,7 +105,6 @@ var optionChangeExcluded = [
     "Scheduler",
     "ScrollView",
     "SlideOut",
-    "SlideOutView",
     "Switch",
     "TileView",
     "Toast",
